move deprecation tests to separate file
